Add alchemy type to alchemy resource hint text
Hint text of this type has an option that can be set to always
display it. This option intentionally can't be set to true without
cheating; it exists so that the game doesn't look up an undefined
property (which would have the correct behavior but seems like a bad
idea).