cleanup info modal
When the credits stuff is merged I may need to figure out a different functionality in order to show the credits as a modal (or however else we want to do it), as right now it's set up to show only links. I can take care of that then, as it'll be easier to think about when we have an actual idea of what we want.