Fix automator bug with UNTIL PRESTIGE
Prestige events are delivered to the commandState of the top stack frame, but
loop bodies execute in their own stack frame. This means that prestige events
only reach the `UNTIL` command during the same automator tick where the loop
condition is being evaluated.

In practice this means that, say, `UNTIL ETERNITY` loops will "miss" any
eternities that happen during execution of one of the commands inside the loop.
This has been observed in bug report
https://discord.com/channels/351476683016241162/1080463168121352202

The fix is to update `commandState` for *each* frame in the stack any time
there is a prestige event. I added a `forEach` to the stack structure so that
we could apply the same logic to all frames. Performance impact should not be
noticeable, assuming that if/while/until are not too deeply nested. 😅

An alternative fix I considered was to copy some `commandState` values from one
frame to another during `AutomatorBackend.nextCommand()`. However, the frame's
`commandState` is completely reset between commands, and leaking state between
them seemed fraught.