Improved Continuum performance
To get if this Dimension is available for purchase, with totalAmount here this would generate a long chain of commands that goes like
(tier - 1).totalAmount -> continuumAmount -> continuumValue -> isAvailableForPurchase -> (...)
for all 8 tiers if we start from AD 8. This only happens if Continuum is enabled, so if we check and return for the Eternity Milestone first, which can be safely assumed it's permanently unlocked at this point of the game, this shouldn't happen.