Add a screen reader mode
This is an option under the visual subtab that anyone with a screen reader can turn on and persists across saves.

Future pull requests by me will use ```
v-if="$viewModel.srMode"
```
to swap out inaccessible parts of the UI with accessible ones on a case-by-case basis, since tabs often have both accessible and inaccessible elements.